labs/week-30-failover-simulation/services/primary-db/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
redis==5.0.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
prometheus-client==0.19.0
pydantic==2.5.3
